requirements.txt
# =============================================================================
# Datapizza Streamlit LLM Interface v1.4.0
# =============================================================================
#
# ⚠️ IMPORTANTE: Ordine di installazione!
# Se pip install -r requirements.txt fallisce, installa manualmente:
#
# 1. pip install streamlit python-dotenv pyyaml reportlab
# 2. pip install datapizza-ai
# 3. pip install datapizza-ai-clients-openai-like
# 4. pip install chromadb beautifulsoup4 PyPDF2 mwclient
#
# =============================================================================

# Core Framework
streamlit>=1.28.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Datapizza AI (installare PRIMA dei clients!)
datapizza-ai

# Datapizza Clients (installare DOPO datapizza-ai!)
# Ollama locale (obbligatorio)
datapizza-ai-clients-openai-like

# Cloud providers (opzionali - decommentare se necessari)
# datapizza-ai-clients-openai
# datapizza-ai-clients-anthropic
# datapizza-ai-clients-google

# Knowledge Base / RAG
chromadb>=0.4.0
beautifulsoup4>=4.12.0
PyPDF2>=3.0.0

# MediaWiki Adapter
mwclient>=0.10.0

# Export PDF
reportlab>=4.0.0
